Added special character recognition
Was a problem that for show "Say "I Love you"" when searching for "I" would not return it. Made it so that when we add shows, we add for each space separated word and also each space separated word with the occlusion of special characters so that we would add to the ADT both an entry for "I" and ""I" and "you" and "you"" etc. This also makes that noticeable change that "S.H.I.E.L.D." show will have two search terms, "shield" and "S.H.I.E.L.D."